Reorganizes Klee source directory
Moves IO files to `io` subdirectory.
