+++ src/axom/klee/KleeError.hpp
@@ -3,13 +3,14 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#ifndef AXOM_KLEEERROR_HPP
-#define AXOM_KLEEERROR_HPP
+#ifndef AXOM_KLEE_ERROR_HPP_
+#define AXOM_KLEE_ERROR_HPP_
+
+
+#include "axom/inlet/inlet_utils.hpp"
 
 #include <exception>
 #include <vector>
-
-#include "axom/inlet/inlet_utils.hpp"
 
 namespace axom
 {
@@ -54,4 +55,4 @@ private:
 }  // namespace klee
 }  // namespace axom
 
-#endif  //AXOM_KLEEERROR_HPP
+#endif  // AXOM_KLEE_ERROR_HPP_
