+++ src/axom/klee/Units.cpp
@@ -3,9 +3,9 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "axom/klee/Units.hpp"
-
 #include "axom/inlet/Proxy.hpp"
+
+#include "axom/klee/Units.hpp"
 #include "axom/klee/KleeError.hpp"
 
 #include <stdexcept>
